requirements.txt
pyyaml>=6.0
jsonbender==0.9.3
pylint
pytest
